test(test-d): add test-d files for vitest typechecking: (#1248)
# Pull Request Checklist

- [x] Did you add new tests and confirm existing tests pass? (`yarn test`)
- [x] Did you update relevant docs? (docs are found in the `site` folder, and guidelines for updating/adding docs can be found in the [contribution guide](https://github.com/alchemyplatform/aa-sdk/blob/main/CONTRIBUTING.md))
- [x] Do your commits follow the [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) standard?
- [x] Does your PR title also follow the [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) standard?
- [x] If you have a breaking change, is it [correctly reflected in your commit message](https://www.conventionalcommits.org/en/v1.0.0/#examples)? (e.g. `feat!: breaking change`)
- [x] Did you run lint (`yarn lint:check`) and fix any issues? (`yarn lint:write`)
- [x] Did you follow the [contribution guidelines](https://github.com/alchemyplatform/aa-sdk/blob/main/CONTRIBUTING.md)?

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on enhancing type checking and testing for the `account-kit` packages, particularly by adding `test-d.ts` files and updating test commands. It also removes unused code related to Alchemy transport.

### Detailed summary
- Added `test-d.ts` files for type definitions in multiple packages.
- Updated `.github/workflows/on-pull-request.yml` to include type checking in tests.
- Modified `package.json` to add a new `test:typecheck` script.
- Updated `CONTRIBUTING.md` to reflect changes in testing commands.
- Removed Alchemy-related code from `client.test.ts`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
