--- site/pages/reference/account-kit/infra/functions/isAlchemyTransport.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: isAlchemyTransport
-description: Overview of the isAlchemyTransport method
----
-
-# isAlchemyTransport
-
-Checks whether the given transport is an AlchemyTransport.
-
-## Import
-
-```ts
-import { isAlchemyTransport } from "@account-kit/infra";
-```
-
-## Usage
-
-```ts
-import { isAlchemyTransport } from "@account-kit/infra";
-if (isAlchemyTransport(transport)) {
-  // do things here that are only when transport is an AlchemyTransport
-}
-```
-
-## Parameters
-
-### x
-
-`Transport | AlchemyTransport`
-The transport to check.
-
-## Returns
-
-`boolean`
-True if the transport is of type AlchemyTransport, otherwise false.
